Further refactor, attempt robustness
