+++ packages/malloy-vscode/src/extension/message_types.ts
@@ -1,3 +1,7 @@
+import { ResultJSON } from "@malloydata/malloy";
+import { DataStyles } from "@malloydata/render";
+import { ConnectionConfig } from "../common";
+
 export enum QueryRunStatus {
   Compiling = "compiling",
   Running = "running",
@@ -10,3 +14,145 @@ export enum QueryMessageType {
   AppReady = "app-ready",
   StartDownload = "start-download",
 }
+
+interface QueryMessageStatusCompiling {
+  type: QueryMessageType.QueryStatus;
+  status: QueryRunStatus.Compiling;
+}
+
+interface QueryMessageStatusRunning {
+  type: QueryMessageType.QueryStatus;
+  status: QueryRunStatus.Running;
+}
+
+interface QueryMessageStatusError {
+  type: QueryMessageType.QueryStatus;
+  status: QueryRunStatus.Error;
+  error: string;
+}
+
+interface QueryMessageStatusDone {
+  type: QueryMessageType.QueryStatus;
+  status: QueryRunStatus.Done;
+  result: ResultJSON;
+  styles: DataStyles;
+}
+
+type QueryMessageStatus =
+  | QueryMessageStatusCompiling
+  | QueryMessageStatusError
+  | QueryMessageStatusRunning
+  | QueryMessageStatusDone;
+
+interface QueryMessageAppReady {
+  type: QueryMessageType.AppReady;
+}
+
+export interface QueryDownloadOptions {
+  format: "json" | "csv";
+  amount: "current" | "all" | number;
+}
+
+interface QueryMessageStartDownload {
+  type: QueryMessageType.StartDownload;
+  downloadOptions: QueryDownloadOptions;
+}
+
+export type QueryPanelMessage =
+  | QueryMessageStatus
+  | QueryMessageAppReady
+  | QueryMessageStartDownload;
+
+export interface WorkerQueryPanelMessage {
+  message: QueryPanelMessage;
+  panelId: string;
+}
+
+export enum ConnectionMessageType {
+  SetConnections = "set-connections",
+  AppReady = "app-ready",
+  TestConnection = "test-connection",
+  RequestBigQueryServiceAccountKeyFile = "request-bigquery-service-account-key-file",
+}
+
+interface ConnectionMessageSetConnections {
+  type: ConnectionMessageType.SetConnections;
+  connections: ConnectionConfig[];
+}
+
+interface ConnectionMessageAppReady {
+  type: ConnectionMessageType.AppReady;
+}
+
+export enum ConnectionTestStatus {
+  Waiting = "waiting",
+  Success = "success",
+  Error = "error",
+}
+
+interface ConnectionMessageTestConnectionWaiting {
+  type: ConnectionMessageType.TestConnection;
+  status: ConnectionTestStatus.Waiting;
+  connection: ConnectionConfig;
+}
+
+interface ConnectionMessageTestConnectionSuccess {
+  type: ConnectionMessageType.TestConnection;
+  status: ConnectionTestStatus.Success;
+  connection: ConnectionConfig;
+}
+
+interface ConnectionMessageTestConnectionError {
+  type: ConnectionMessageType.TestConnection;
+  status: ConnectionTestStatus.Error;
+  error: string;
+  connection: ConnectionConfig;
+}
+
+export type ConnectionMessageTest =
+  | ConnectionMessageTestConnectionWaiting
+  | ConnectionMessageTestConnectionSuccess
+  | ConnectionMessageTestConnectionError;
+
+export enum ConnectionServiceAccountKeyRequestStatus {
+  Waiting = "waiting",
+  Success = "success",
+}
+
+interface ConnectionMessageServiceAccountKeyRequestWaiting {
+  type: ConnectionMessageType.RequestBigQueryServiceAccountKeyFile;
+  status: ConnectionServiceAccountKeyRequestStatus.Waiting;
+  connectionId: string;
+}
+
+interface ConnectionMessageServiceAccountKeyRequestSuccess {
+  type: ConnectionMessageType.RequestBigQueryServiceAccountKeyFile;
+  status: ConnectionServiceAccountKeyRequestStatus.Success;
+  connectionId: string;
+  serviceAccountKeyPath: string;
+}
+
+export type ConnectionMessageServiceAccountKeyRequest =
+  | ConnectionMessageServiceAccountKeyRequestWaiting
+  | ConnectionMessageServiceAccountKeyRequestSuccess;
+
+export type ConnectionPanelMessage =
+  | ConnectionMessageAppReady
+  | ConnectionMessageSetConnections
+  | ConnectionMessageTest
+  | ConnectionMessageServiceAccountKeyRequest;
+
+export enum HelpMessageType {
+  AppReady = "app-ready",
+  EditConnections = "edit-connections",
+}
+
+interface HelpMessageAppReady {
+  type: HelpMessageType.AppReady;
+}
+
+interface HelpMessageEditConnections {
+  type: HelpMessageType.EditConnections;
+}
+
+export type HelpPanelMessage = HelpMessageAppReady | HelpMessageEditConnections;
