+++ packages/malloy-vscode/src/extension/state.ts
@@ -13,10 +13,8 @@
 
 import { TextDocument, WebviewPanel } from "vscode";
 import { Result } from "@malloydata/malloy";
-import {
-  QueryPanelMessage,
-  WebviewMessageManager,
-} from "./webview_message_manager";
+import { QueryPanelMessage } from "./message_types";
+import { WebviewMessageManager } from "./webview_message_manager";
 import { VSCodeConnectionManager } from "./connection_manager";
 
 export const CONNECTION_MANAGER = new VSCodeConnectionManager();
