+++ packages/malloy-vscode/src/extension/webviews/connections_page/connections_vscode_context.ts
@@ -11,7 +11,7 @@
  * GNU General Public License for more details.
  */
 
-import { ConnectionPanelMessage } from "../../webview_message_manager";
+import { ConnectionPanelMessage } from "../../message_types";
 import { makeVSCodeContext } from "../vscode_context";
 import { makeUseVSCodeContext } from "../vscode_context";
