+++ packages/malloy-vscode/src/extension/webviews/query_page/entry.ts
@@ -16,7 +16,7 @@ import * as _webviewAPI from "vscode-webview";
 import ReactDOM from "react-dom";
 import React from "react";
 import { App } from "./App";
-import { QueryPanelMessage } from "../../webview_message_manager";
+import { QueryPanelMessage } from "../../message_types";
 import { getVSCodeAPI, QueryVSCodeContext } from "./query_vscode_context";
 
 (() => {
